requirements.txt
pyOpenssl
Flask-RESTful==0.3.9
Flask==2.3.0
Flask-Cors==3.0.10
gunicorn==20.1.0